Remove rb_validation from utils and factories
